+++ VersionHistory.md
@@ -10,6 +10,8 @@ Prefix the description of the change with `[major]`, `[minor]` or `[patch]` in a
 * Change .NET Framework minimum version to 4.6.1.
 * [major] Remove `DefaultValueDefaultAttribute`.
 * [major] Remove `JsonPointer` and `JsonPatch`.
+* [major] Simplify `JsonInputSettings` and `JsonOutputSettings`.
+* [major] Rename `JsonFilter.AlternatePathSepartor` to `AlternatePathSeparator`.
 
 ## Released